optimize token page loading speed
